Move setup logic into main.
