+++ package.json
@@ -67,7 +67,8 @@
     "@actions/attest": "^1.5.0",
     "@actions/cache": "^3.3.0",
     "@actions/core": "^1.11.1",
-    "@actions/github": "^6.0.0"
+    "@actions/github": "^6.0.0",
+    "@actions/tool-cache": "^2.0.1"
   },
   "devDependencies": {
     "@babel/core": "^7.26.0",
