+++ src/main.js
@@ -1,6 +1,5 @@
 const core = require('@actions/core')
 const tc = require('@actions/tool-cache')
-const { setup } = require('./setup')
 
 /**
  * The main function for the action.
@@ -8,11 +7,50 @@ const { setup } = require('./setup')
  */
 async function run() {
   try {
-    core.debug(`Setup Nexus IQ CLI`)
-    await setup()
+    // Download URL for the Nexus IQ CLI
+    const cliDownloadURLTenplate =
+      'https://download.sonatype.com/clm/scanner/nexus-iq-cli-{iq-cli-version}.jar'
+    // Get version of tool to be installed
+    const version = core.getInput('version')
+
+    // Throw an error if the version is not provided
+    if (!version) {
+      throw new Error('Input parameter version is required.')
+    }
+
+    // Replace the version placeholder in the download URL
+    const cliDownloadURL = cliDownloadURLTenplate.replace(
+      '{iq-cli-version}',
+      version
+    )
+
+    // File chache path for the tool
+    let cachedToolPath = tc.find('iq-cli', version)
+
+    // If the tool is not already cached, download it
+    if (!cachedToolPath) {
+      core.debug(`Downloading IQ CLI version ${version}`)
+      const toolPath = await tc.downloadTool(
+        cliDownloadURL,
+        'sonatype-iq-cli.jar'
+      )
+      core.debug(`Downloaded IQ CLI to ${toolPath}`)
+      cachedToolPath = await tc.cacheFile(
+        toolPath,
+        'sonatype-iq-cli.jar',
+        'iq-cli',
+        version
+      )
+    } else {
+      core.debug(`IQ CLI version ${version} found in cache.`)
+    }
+
+    core.setOutput('iq-cli-version', version)
   } catch (error) {
     // Fail the workflow run if an error occurs
-    core.setFailed(error.message)
+    if (error instanceof Error) {
+      core.setFailed(error.message)
+    }
   }
 }
